fix: Restructuring the project folder structure
